+++ src/features/optional/major/fastTravel/trapdoor.ts
@@ -122,7 +122,7 @@ function shouldRemove() {
     g.race.myStatus === RacerStatus.RACING &&
     g.race.goal === RaceGoal.HUSH &&
     stage === LevelStage.WOMB_2 &&
-    roomGridIndex !== GridRoom.BLUE_WOMB // eslint-disable-line isaacscript/strict-enums
+    roomGridIndex !== (GridRoom.BLUE_WOMB as int)
   ) {
     log(
       `Removed a vanilla trapdoor after Mom on game frame: ${gameFrameCount}`,
@@ -136,7 +136,7 @@ function shouldRemove() {
     g.race.myStatus === RacerStatus.RACING &&
     g.race.goal === RaceGoal.HUSH &&
     stage === LevelStage.BLUE_WOMB &&
-    roomGridIndex !== GridRoom.THE_VOID // eslint-disable-line isaacscript/strict-enums
+    roomGridIndex !== (GridRoom.THE_VOID as int)
   ) {
     log(
       `Removed a vanilla trapdoor after Hush (for a Hush goal) on game frame: ${gameFrameCount}`,
@@ -193,7 +193,7 @@ function shouldRemove() {
   if (
     stage < LevelStage.WOMB_1 &&
     backwardPath &&
-    roomGridIndex !== GridRoom.SECRET_EXIT // eslint-disable-line isaacscript/strict-enums
+    roomGridIndex !== (GridRoom.SECRET_EXIT as int)
   ) {
     log(
       `Removed a vanilla trapdoor on the Ascent on game frame: ${gameFrameCount}`,
